Add files via upload
This file is for you to run both conversion of the trajectories into xtc and to concatenate them. NOTE: you will need to run prepfiles.py first! and you will need to tweak the code if instead of doing step7.trr you are doing step5.xtc. It is still necessary to correct for the PBC.